Add CMake build system
This commit adds a complete CMake-based build system targetting Windows,
Mac OS X and UNIX-likes.

Whenever a non-critical library cannot be found, the dependant features
are simply disabled and MilkyTracker can be built without them.

System-provided libraries are used for the build instead of the bundled
versions, except on Windows, which still requires the library sources.

On Windows and OS X, all libraries not provided by the operating system
are linked statically. This includes the Visual C/C++ runtime on
Windows (alleviating the need to install a runtime).