CMake: Add CPack packaging capability
This commit adds the ability to build binary packages, which is mainly
useful under OS X and Windows.

Under OS X, a DMG drag ’n drop disk image is built with a customised
Finder backdrop, containing the app, documentation and sample songs.
The custom backdrop consists of a standard resolution version and a
Retina resolution version. These were combined together into a
multi-representation TIFF using tiffutil - the TIFF is what gets copied
to the DMG, and the OS decides which resolution to use when displaying
it.

Under Windows, a ZIP file is created, also with relevant documentation
and sample songs.

For the Unix-like OSes, the program and accompanying resources are
copied to their appropriate destinations underneath the prefix (set by
the traditional DESTDIR variable).