PatternEditor: Use eKeyChar for data input events
Previously, input into the instrument, volume, and effects columns relied on
eKeyDown supplying an ASCII character value in the { VK, SC, char } array coming
in through the event queue.

Whilst this works most of the time as far as entering alphanumberical values is
concerned (i.e. hex values in the ins/effects columns, and 'D', 'L', 'M', 'P'
etc. in the volume column), entering '+' was not possible on some keyboard
layouts in the SDL version because '+' is a shifted character on some keyboards
(e.g. UK), and SDL separates the notion of text input from physical keystrokes.

To increase reliability when using international keyboard layouts, we have the
PatternEditorControl additionally listen for eKeyChar (text input) events, and
act accordingly when characters of interest are received.