Eliminate all remaining compiler warnings
The majority of warnings were for switch statements with unused cases
and assignments inside conditionals, which required additional
parentheses.  However, some of the warnings led to the discovery of
bugs, for instance base classes with no virtual deconstructor - which
would result in a memory leak when deleting the derived classes.